Add support for different balena boot partitions, + jetson nano
This PR needs to do a few things to add support for nvidia jetson devices, and in particular the jetson nano.

Besides adding a new machine type, it needs to handle the complex partition layout on the jetson devices. The approach used should be generic to all balena devices, and has been tested on the Jetson Nano, Jetson Xavier NX, and Raspberry Pi 4.

To allow OTA updates of the host image, the nano has a total of 16 partitions, and we need to iterate across them to find the correct boot partition named resin-boot. The fatfs library doesn't include a native way to do this, so instead we loop over the 128 possible GPT partitions for completeness. It doesn't take an unreasonable amount of time, and several conditions allow exiting early. It is of particular importance to "close" each partition reference after it's been opened, as the library does not handle multiple access or partitions well, giving inexplicable errors if this is not done.

As the fat library we use doesn't support directly specifying a partition  for the `cp` method, instead a reference is passed to the copy_in and copy_out methods.